+++ CHANGELOG.md
@@ -9,6 +9,12 @@
 - Define open ai model in env var
 - Input context pompts
 
-## [0.1.3] - 2023-03-23
+## [0.1.3] - 2023-03-25
 
 - Custom contexts prompts support
+
+## [0.1.4] - 2023-03-25
+
+- Added messages history
+- Added more context prompts
+- Added Dall-E support